add cmd arguments

+++ lib/cardmaker.rb
@@ -3,12 +3,12 @@ require 'fileutils'
 require_relative 'read_worksheet'
 require_relative 'render_card'
 require_relative 'symbol_replace'
+require_relative 'define_options'
 
-if (ARGV.empty?)
-  raise "Usage: (ruby) cardmaker(.rb) [GOOGLE DRIVE FILENAME]\n #{ARGV.length}"
-end
+options = {}
+DefineOptions.new(options)
 
-docname = ARGV[0].freeze
+docname = options['name']
 confpath = File.expand_path('../config/config.json', File.dirname(__FILE__))
 session = GoogleDrive::Session.from_config(confpath)
 
@@ -21,27 +21,27 @@ file.worksheets.each do |ws|
   cards = cards.merge(symbol_replace(read_worksheet(ws)))
 end
 
-r = CardRenderer.new(
-  '/colors.yaml',
-  '/cardlayout.yaml',
-  '/cardlist.yaml',
-  '/symbols.yaml',
-  'images/',
-  '../output/'
-)
+r = CardRenderer.new(options)
 
-FileUtils.rm_rf(Dir['./output/*'])
+if options['clean']
+  FileUtils.rm_rf(Dir['./output/*'])
+end
 
-#cards.each do |title, card|
-card = cards['Troops']
-  #print "\n#{title} ======\n"
+if (options['sheets'])
+  cards.select! { |title, card| options['sheets'].include?(title) }
+end
+
+cards.each do |title, card|
+  print "\n#{title} ======\n" if options['verbose']
 
   card.each do |t, c|
-    print "#{t}\n"
+    print "#{t}\n" if options['verbose']
 
     r.render_card(c)
   end
-#end
+end
 
-Dir.mkdir('./output/output')
-#r.render_cardlist('/output/output.png')
+if options['print']
+  FileUtils.mkdir_p('./output/output')
+  r.render_cardlist('/output/output.png')
+end
